Use cursor-based pagination instead of offset pagination
Using cursor-based pagination allows us to more easily find the word of a day for a specific date. Or to find the last X words of the day up to a given date.

This is an infinite pagination, and there is one word for each date in time. We don't need to provide information in the response about the next or previous ids.